Fixed unit tests
