+++ SecretSanta/src/SecretSanta.Data/SecretSanta.Data.csproj
@@ -2,7 +2,6 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-    <nullable>enable</nullable>
   </PropertyGroup>
 
   <ItemGroup>
